Basic Coding Test - Q6 Even Remainder
Problem Description

Write a program to input an integer A and another array B of N numbers from user and print the count of elements of B such that B[i]%A is even.



Problem Constraints

1 <= A <= 100

1 <= N <= 100

1 <= B[i] <= 1000



Input Format

First line is an integer A as per the question.

Second line is N which means number of elements in array B.

Next line contains N space separated integers which are the elements of array B.



Output Format

An integer representing the count of elements such that B[i]%A is even.



Example Input

Input 1:

3
7
2 3 4 5 6 7 8


Example Output

Output 1:

5


Example Explanation

Explanation 1:

For x = 2,5 and 8, x%3=2 which is even. 
Also, for x = 3 and 6, x%3=0 which is even.